+++ src/main/java/com/commercetools/sync/products/utils/ProductSyncUtils.java
@@ -197,4 +197,7 @@ public final class ProductSyncUtils {
                                     .map(Optional::get)
                                     .collect(Collectors.toList());
     }
+
+    private ProductSyncUtils() {
+    }
 }