+++ src/main/java/com/commercetools/sync/products/utils/ProductUpdateActionUtils.java
@@ -711,4 +711,7 @@ public final class ProductUpdateActionUtils {
                 + "Reason: %s",
             product.getKey(), reason));
     }
+
+    private ProductUpdateActionUtils() {
+    }
 }