+++ src/main/java/com/commercetools/sync/products/utils/ProductVariantAttributeUpdateActionUtils.java
@@ -83,4 +83,7 @@ public final class ProductVariantAttributeUpdateActionUtils {
         }
 
     }
+
+    private ProductVariantAttributeUpdateActionUtils() {
+    }
 }